refactor(stella-serve): split remote.rs tests into a sibling submodule
The #3286 witnesses pushed remote.rs over the 1500-line gate; the tests
move to remote/tests.rs (the driver/settlement.rs pattern), no behavior
change.

Refs #3286
